[BUGFIX] Avoid double-encoding when converting to array
The original fix was not the right one. The double encoding is due
to SimpleXML already handling the encoding. The proposed solution
is to add a note in the manual advising to avoid using the encoding
parameter when the target format is an array.

The change to the encoding parameter in the sample configuration
is reverted.

Resolves: #57531
Releases: 1.5
Change-Id: I73e583fca9b9be3bfc900a4f387111c74a8dc24e
Reviewed-on: https://review.typo3.org/29549
Reviewed-by: Francois Suter
Tested-by: Francois Suter